requirements.txt
telethon==1.28.5
pyTelegramBotAPI==4.11.0
python-dotenv==1.0.0